+++ package.json
@@ -16,6 +16,7 @@
     "react": "17.0.1",
     "react-native": "0.64.0",
     "react-native-dev-menu": "^4.0.2",
+    "react-native-svg": "^12.1.1",
     "react-native-svg-transformer": "^0.14.3",
     "react-native-svg-uri": "^1.2.3"
   },
